requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
alabaster==0.7.12
aniso8601==9.0.1
antlr4-python3-runtime==4.8
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.1.0
audioread==3.0.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==19.10b0
boto3==1.26.26
botocore==1.29.26
braceexpand==0.1.7
cachetools==5.2.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.2
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.4
decorator==5.1.1
deepmultilingualpunctuation==1.0.1
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.19
editdistance==0.6.1
einops==0.6.0
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.2.0
faiss-cpu==1.7.3
fasttext==0.9.2
filelock==3.8.0
Flask==2.2.2
Flask-RESTful==0.3.9
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==6.1.1
g2p-en==2.1.0
gdown==4.6.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
h5py==3.7.0
huggingface-hub==0.11.0
hydra-core==1.1.2
HyperPyYAML==1.1.0
idna==3.4
ijson==3.1.4
imagesize==1.4.1
inflect==6.0.2
iniconfig==1.1.1
ipadic==1.0.0
ipykernel==6.19.2
ipython==8.7.0
ipywidgets==8.0.3
isort==4.3.21
itsdangerous==2.1.2
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jiwer==2.5.1
jmespath==1.0.1
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyterlab-widgets==3.0.4
kaldi-python-io==1.2.2
kaldiio==2.17.2
kiwisolver==1.4.4
latexcodec==2.0.1
Levenshtein==0.20.2
librosa==0.9.2
llvmlite==0.39.1
loguru==0.6.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.19.0
matplotlib==3.6.2
matplotlib-inline==0.1.6
mecab-python3==1.0.5
mpmath==1.2.1
multidict==6.0.3
natsort==8.2.0
nemo-toolkit==1.13.0
nest-asyncio==1.5.6
nltk==3.7
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.1.2
onnx==1.12.0
OpenCC==1.1.6
packaging==21.3
pandas==1.5.1
pangu==4.0.6.1
parameterized==0.8.1
parso==0.8.3
pathspec==0.10.2
pathtools==0.1.2
pesq==0.0.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip-api==0.0.30
pipreqs==0.4.11
plac==1.3.5
platformdirs==2.6.0
pluggy==1.0.0
pooch==1.6.0
portalocker==2.6.0
progress==1.6
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.1
pybtex==0.24.0
pybtex-docutils==1.0.2
pycparser==2.21
pydantic==1.10.2
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.13.0
pyparsing==3.0.9
pypinyin==0.47.1
PySocks==1.7.1
pystoi==0.3.3
pytest==7.2.0
pytest-runner==6.0.0
python-dateutil==2.8.2
pytorch-lightning==1.7.7
pytz==2022.6
PyYAML==5.4.1
pyzmq==24.0.1
rapidfuzz==2.13.4
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sacrebleu==2.3.1
sacremoses==0.0.53
scikit-learn==1.1.3
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.11.1
setproctitle==1.3.2
shellingham==1.5.0
shortuuid==1.0.11
simplejson==3.18.0
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.11.0
soupsieve==2.3.2.post1
sox==1.4.1
speechbrain==0.5.13
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-bibtex==2.5.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.1
text-unidecode==1.3
textdistance==4.5.0
texterrors==0.4.4
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
torch==1.13.0
torchaudio==0.13.0
torchmetrics==0.10.3
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
traitlets==5.7.0
transformers==4.21.2
typed-ast==1.5.4
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.12
wandb==0.13.6
wcwidth==0.2.5
webdataset==0.1.62
Werkzeug==2.2.2
wget==3.2
widgetsnbextension==4.0.4
wrapt==1.14.1
yarg==0.1.9
yarl==1.8.2
youtokentome==1.0.6

